+++ pypsa/networks.py
@@ -17,6 +17,7 @@ from pypsa.common import equals, future_deprecation
 from pypsa.components.abstract import Components
 from pypsa.components.common import as_components
 from pypsa.constants import DEFAULT_EPSG, DEFAULT_TIMESTAMP
+from pypsa.statistics.abstract import AbstractStatisticsAccessor
 
 try:
     from cloudpathlib import AnyPath as Path
@@ -87,6 +88,7 @@ from pypsa.plot.accessor import PlotAccessor
 from pypsa.plot.maps import explore, iplot
 from pypsa.statistics import StatisticsAccessor
 from pypsa.typing import is_1d_list_like
+from pypsa.version import __version_semver__
 
 if TYPE_CHECKING:
     import linopy
@@ -288,12 +290,10 @@ class Network:
         # Initialise root logger and set its level, if this has not been done before
         logging.basicConfig(level=logging.INFO)
 
-        from pypsa import release_version as pypsa_version
-
         self.name: str = name
 
         # this will be saved on export
-        self.pypsa_version: str = pypsa_version
+        self.pypsa_version: str = __version_semver__
 
         self._meta: dict = {}
 
@@ -368,19 +368,82 @@ class Network:
         self.merge(other)
 
     def __eq__(self, other: Any) -> bool:
-        """Check for equality of two networks."""
+        """
+        Check for equality of two networks.
+
+        Parameters
+        ----------
+        other : Any
+            The other network to compare with.
+
+        Returns
+        -------
+        bool
+            True if the networks are equal, False otherwise.
+
+        See Also
+        --------
+        pypsa.Network.equals : Check for equality of two networks.
+        """
+        return self.equals(other)
+
+    def equals(self, other: Any, log_mode: str = "silent") -> bool:
+        """
+        Check for equality of two networks.
+
+        Parameters
+        ----------
+        other : Any
+            The other network to compare with.
+        log_mode: str, default="silent"
+            Controls how differences are reported:
+            - 'silent': No logging, just returns True/False
+            - 'verbose': Prints differences but doesn't raise errors
+            - 'strict': Raises ValueError on first difference
+
+        Raises
+        ------
+        ValueError
+            If log_mode is 'strict' and components are not equal.
+
+        Returns
+        -------
+        bool
+            True if the networks are equal, False otherwise.
+
+        Examples
+        --------
+        >>> n1 = pypsa.Network()
+        >>> n2 = pypsa.Network()
+        >>> n1.add("Bus", "bus1")
+        Index(['bus1'], dtype='object')
+        >>> n2.add("Bus", "bus2")
+        Index(['bus2'], dtype='object')
+        >>> n1.equals(n2)
+        False
+
+        """
         ignore = [
             OptimizationAccessor,
             ClusteringAccessor,
             StatisticsAccessor,
             PlotAccessor,
+            AbstractStatisticsAccessor,
         ]
-
+        not_equal = False
         if isinstance(other, self.__class__):
             for key, value in self.__dict__.items():
-                if not equals(value, other.__dict__[key], ignored_classes=ignore):
+                if not equals(
+                    value,
+                    other.__dict__[key],
+                    ignored_classes=ignore,
+                    log_mode=log_mode,
+                    path="n." + key,
+                ):
                     logger.warning("Mismatch in attribute: %s", key)
-                    return False
+                    not_equal = True
+                    if not log_mode:
+                        break
         else:
             logger.warning(
                 "Can only compare two pypsa.Network objects with each other. Got %s.",
@@ -388,7 +451,8 @@ class Network:
             )
 
             return False
-        return True
+
+        return not not_equal
 
     # ----------------
     # Initialization